Moves semantic validation to separate class/object
