+++ benchflow-dsl/src/main/scala/cloud/benchflow/dsl/BenchFlowTestAPI.scala
@@ -3,7 +3,7 @@ package cloud.benchflow.dsl
 import cloud.benchflow.dsl.definition.BenchFlowTest
 import cloud.benchflow.dsl.definition.BenchFlowTestYamlProtocol._
 import cloud.benchflow.dsl.definition.errorhandling.{ BenchFlowDeserializationException, BenchFlowDeserializationExceptionMessage }
-import cloud.benchflow.dsl.definition.configuration.goal.goaltype.GoalType.LOAD
+import cloud.benchflow.dsl.definition.validation.SemanticValidation
 import net.jcazevedo.moultingyaml._
 
 import scala.util.{ Failure, Success, Try }
@@ -38,8 +38,7 @@ object BenchFlowTestAPI {
     }
 
     // validate semantic in separate function on the object
-    val validationErrorMessages = new StringBuilder
-    val isValid = validateTest(test, validationErrorMessages)
+    val (isValid, validationErrorMessages) = SemanticValidation.validateTest(test)
 
     if (!isValid) {
       throw BenchFlowDeserializationExceptionMessage(validationErrorMessages.toString())
@@ -56,88 +55,6 @@ object BenchFlowTestAPI {
     val testYaml: YamlObject = benchFlowTest.toYaml.asYamlObject
 
     testYaml.prettyPrint
-
-  }
-
-  private val userMissing = "Missing user specification. "
-
-  def validateTest(benchFlowTest: BenchFlowTest, validationErrorMessages: StringBuilder): Boolean = {
-
-    validationErrorMessages.append("Semantic validation failed. ")
-    var isValid = true
-
-    // TODO - validate semantic in separate function on the object
-
-    // check that number of users is defined if exploration goal
-    isValid = ifExplorationThenUsers(benchFlowTest, validationErrorMessages)
-    // check if load that users are specified
-    isValid = ifLoadGoalThenUsers(benchFlowTest, validationErrorMessages) && isValid
-    // check that exploration space is defined if such goal
-    isValid = ifExplorationThenExplorationSpace(benchFlowTest, validationErrorMessages) && isValid
-    // check that at least one workload is defined
-    isValid = workloadIsDefined(benchFlowTest, validationErrorMessages) && isValid
-
-    isValid
-  }
-
-  private def ifExplorationThenUsers(benchFlowTest: BenchFlowTest, validationErrorMessages: StringBuilder): Boolean = {
-
-    var isValid = true
-
-    if (benchFlowTest.configuration.goal.goalType != LOAD && benchFlowTest.configuration.users.isEmpty) {
-      benchFlowTest.configuration.goal.explorationSpace match {
-        case Some(explorationSpace) => explorationSpace.workload match {
-          case Some(workload) => if (workload.users.isEmpty) {
-            isValid = false
-            validationErrorMessages.append(userMissing)
-          }
-          case None =>
-            isValid = false
-            validationErrorMessages.append(userMissing)
-        }
-        case None =>
-          isValid = false
-          validationErrorMessages.append(userMissing)
-      }
-    }
-
-    isValid
-  }
-
-  private def ifLoadGoalThenUsers(benchFlowTest: BenchFlowTest, validationErrorMessages: StringBuilder): Boolean = {
-    var isValid = true
-
-    if (benchFlowTest.configuration.goal.goalType == LOAD && benchFlowTest.configuration.users.isEmpty) {
-      isValid = false
-      validationErrorMessages.append(userMissing)
-    }
-
-    isValid
-
-  }
-
-  def ifExplorationThenExplorationSpace(benchFlowTest: BenchFlowTest, validationErrorMessages: StringBuilder): Boolean = {
-
-    var isValid = true
-
-    if (benchFlowTest.configuration.goal.goalType != LOAD && benchFlowTest.configuration.goal.explorationSpace.isEmpty) {
-      isValid = false
-      validationErrorMessages.append("Missing exploration space specification. ")
-    }
-
-    isValid
-  }
-
-  def workloadIsDefined(benchFlowTest: BenchFlowTest, validationErrorMessages: StringBuilder): Boolean = {
-
-    var isValid = true
-
-    if (benchFlowTest.workload.isEmpty) {
-      isValid = false
-      validationErrorMessages.append("Missing workload specification. ")
-    }
-
-    isValid
 
   }
